Organized files in MVVM
