+++ Travel-Planner/Travel-Planner/Screens/ItineraryView/ItineraryView.swift
@@ -25,7 +25,7 @@ struct ItineraryView: View {
         NavigationView{
             VStack{
                 
-                HScollDatesView(dayPlans: itinerary.dayPlans,
+                HScrollDatesView(dayPlans: itinerary.dayPlans,
                                 selectedDate: $selectedDates)
                 
                 Spacer()
@@ -42,48 +42,3 @@ struct ItineraryView: View {
     ItineraryView()
 }
 
-struct DateCapsuleView: View {
-    
-    var date: Date
-    var isSelected: Bool
-    
-    var body: some View {
-        Text(DateUtilities.formattedDate(date))
-            .font(.system(size: 18))
-            .fontWeight(.semibold)
-            .frame(width: 145, height: 35)
-            .background(isSelected ? Color.brandPrimary : Color.gray)
-            .foregroundColor(.white)
-            .cornerRadius(100)
-            .padding(3)
-    }
-}
-
-struct HScollDatesView: View {
-    
-    var dayPlans: [DayPlan]
-    @Binding var selectedDate: Date?
-    
-    var body: some View {
-        ScrollView(.horizontal) {
-            LazyHStack {
-                ForEach(dayPlans){ dayPlan in
-                    DateCapsuleView(date: dayPlan.date,
-                                    isSelected: isSelected(date: dayPlan.date))
-                    .onTapGesture {
-                        selectedDate = dayPlan.date
-                    }
-                }
-            }
-        }
-        .frame(height: 80)
-    }
-    
-    func isSelected(date: Date) -> Bool {
-        if let selectedDate = self.selectedDate {
-            return Calendar.current.isDate(selectedDate, inSameDayAs: date)
-        } else {
-            return false
-        }
-    }
-}
