+++ Travel-Planner/Travel-Planner/Screens/TravelPlannerView.swift
@@ -9,6 +9,7 @@ import SwiftUI
 
 struct TravelPlannerView: View {
     var body: some View {
+        Text("This is the Itinerary List View")
         ItineraryView()
     }
 }
